refactor: extract window.c from somewm.c
Move window lifecycle, geometry, and client commands into a dedicated
module. XDG shell handlers (create/map/unmap/destroy), popup and
decoration handling, geometry application (apply_geometry_to_wlroots,
resize, applybounds), layout entry (arrange), appearance helpers,
client commands (setmon, setfullscreen, swapstack, zoom, etc.),
and hot-reload listener management.

35 functions, 3 listener structs, and the Popup typedef extracted.
Client.c functions stay in objects/client.c for now - that split
happens naturally during Phase 2 (event queue) when the binding
layer and engine layer must separate.

somewm.c: 3471 -> 1811 lines (-1660)
window.c: 1744 lines (new)
